+++ sgl-model-gateway/src/server.rs
@@ -68,102 +68,14 @@ async fn parse_function_call(
     State(state): State<Arc<AppState>>,
     Json(req): Json<ParseFunctionCallRequest>,
 ) -> Response {
-    match &state.context.tool_parser_factory {
-        Some(factory) => match factory.registry().get_pooled_parser(&req.tool_call_parser) {
-            Some(pooled_parser) => {
-                let parser = pooled_parser.lock().await;
-                match parser.parse_complete(&req.text).await {
-                    Ok((remaining_text, tool_calls)) => (
-                        StatusCode::OK,
-                        Json(json!({
-                            "remaining_text": remaining_text,
-                            "tool_calls": tool_calls,
-                            "success": true
-                        })),
-                    )
-                        .into_response(),
-                    Err(e) => {
-                        error!("Failed to parse function calls: {}", e);
-                        (
-                            StatusCode::BAD_REQUEST,
-                            Json(json!({
-                                "error": format!("Failed to parse function calls: {}", e),
-                                "success": false
-                            })),
-                        )
-                            .into_response()
-                    }
-                }
-            }
-            None => (
-                StatusCode::BAD_REQUEST,
-                Json(json!({
-                    "error": format!("Unknown tool parser: {}", req.tool_call_parser),
-                    "success": false
-                })),
-            )
-                .into_response(),
-        },
-        None => (
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({
-                "error": "Tool parser factory not initialized",
-                "success": false
-            })),
-        )
-            .into_response(),
-    }
+    state.router.parse_function_call(&req).await
 }
 
 async fn parse_reasoning(
     State(state): State<Arc<AppState>>,
     Json(req): Json<SeparateReasoningRequest>,
 ) -> Response {
-    match &state.context.reasoning_parser_factory {
-        Some(factory) => match factory.registry().get_pooled_parser(&req.reasoning_parser) {
-            Some(pooled_parser) => {
-                let mut parser = pooled_parser.lock().await;
-                match parser.detect_and_parse_reasoning(&req.text) {
-                    Ok(result) => (
-                        StatusCode::OK,
-                        Json(json!({
-                            "normal_text": result.normal_text,
-                            "reasoning_text": result.reasoning_text,
-                            "success": true
-                        })),
-                    )
-                        .into_response(),
-                    Err(e) => {
-                        error!("Failed to separate reasoning: {}", e);
-                        (
-                            StatusCode::BAD_REQUEST,
-                            Json(json!({
-                                "error": format!("Failed to separate reasoning: {}", e),
-                                "success": false
-                            })),
-                        )
-                            .into_response()
-                    }
-                }
-            }
-            None => (
-                StatusCode::BAD_REQUEST,
-                Json(json!({
-                    "error": format!("Unknown reasoning parser: {}", req.reasoning_parser),
-                    "success": false
-                })),
-            )
-                .into_response(),
-        },
-        None => (
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({
-                "error": "Reasoning parser factory not initialized",
-                "success": false
-            })),
-        )
-            .into_response(),
-    }
+    state.router.parse_reasoning(&req).await
 }
 
 async fn sink_handler() -> Response {
